+++ client/src/components/login/MicrosoftSignInButton.tsx
@@ -1,0 +1,24 @@
+import type { Component, JSX } from 'solid-js'
+import msLogo from '../../img/ms-symbollockup_mssymbol_19.svg'
+import '../../css/microsoft-login.css'
+
+type Props = {
+	localLoginForm: JSX.Element
+}
+
+const MicrosoftSignInButton: Component<Props> = (props) => (
+	<div class="microsoft-login-section">
+		<a href="/api/auth/login/microsoft" class="login-button-microsoft">
+			<span class="ms-logo" aria-hidden="true">
+				<img src={msLogo} alt="Microsoft logo" />
+			</span>
+			<span class="ms-text">Sign in with Microsoft</span>
+		</a>
+		<details class="local-login-details">
+			<summary>Use local account</summary>
+			{props.localLoginForm}
+		</details>
+	</div>
+)
+
+export default MicrosoftSignInButton
