+++ client/src/css/login.css
@@ -1,16 +1,18 @@
 .login-container {
+	min-height: 100%;
 	display: flex;
+
+	/* horizontal */
 	justify-content: center;
-	align-items: center;
-	min-height: 100vh;
-	background-color: var(--color-bg-muted);
+
+	/* vertical */
+	align-items: flex-start;
+	padding: clamp(1rem, 20vh, 6rem) 1rem 2rem;
 }
 
 .login-card {
 	background: var(--color-bg);
 	padding: 2rem;
-	border-radius: var(--corner-radius);
-	box-shadow: var(--shadow-card);
 	width: 100%;
 	max-width: 400px;
 }
